changed word frequency processing

+++ TagCloud/WordStatistics/IWordStatistics.cs
@@ -2,7 +2,7 @@ namespace TagCloud.WordStatistics;
 
 public interface IWordStatistics
 {
-    public double GetWordFrequency(string word);
+    public float GetWordFrequency(string word);
     public IEnumerable<string> GetWords();
     public void Populate(IEnumerable<string> words);
 }
